requirements.txt
# PyTorch and Deep Learning - Compatible with Mac M4 and Google Colab
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Data Science and ML
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Computer Vision
opencv-python>=4.5.0
Pillow>=8.3.0
albumentations>=1.3.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Progress bars and utilities
tqdm>=4.62.0
rich>=12.0.0

# Jupyter and Notebook support
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=7.6.0

# Dataset handling
datasets>=2.0.0
huggingface-hub>=0.14.0

# Configuration management
pyyaml>=6.0.0
python-dotenv>=0.19.0

# Experiment tracking (optional)
wandb>=0.13.0
tensorboard>=2.8.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Code quality
black>=22.0.0
flake8>=5.0.0
isort>=5.10.0

# Development tools
pre-commit>=2.17.0

# Platform-specific optimizations
# For Mac M4: These will automatically use Apple Silicon optimizations when available
# For Colab: These will use CUDA when available